requirements.txt
loguru
mkdocs
pip
python-dotenv
ruff
tqdm
typer
cookiecutter-data-science
pandas
numpy
scikit-learn
ipykernel
ipywidgets
torch
matplotlib
seaborn
xgboost
imblearn
lifelines
shap